Changes from CODE REVIEW

+++ spec/lib/carto/ghost_tables_manager_spec.rb
@@ -229,35 +229,34 @@ module Carto
     end
 
     it 'should link raster tables' do
-      if ::User[@user.id].in_database.table_exists?('raster_overviews')
-        run_in_user_database(%{
-          CREATE TABLE manolo_raster ("cartodb_id" uuid, "the_raster_webmercator" raster);
-          CREATE TRIGGER test_quota_per_row
-            BEFORE INSERT OR UPDATE
-            ON manolo_raster
-            FOR EACH ROW
-            EXECUTE PROCEDURE cdb_checkquota('0.001', '-1', 'public');
-        })
+      next unless ::User[@user.id].in_database.table_exists?('raster_overviews')
+      run_in_user_database(%{
+        CREATE TABLE manolo_raster ("cartodb_id" uuid, "the_raster_webmercator" raster);
+        CREATE TRIGGER test_quota_per_row
+          BEFORE INSERT OR UPDATE
+          ON manolo_raster
+          FOR EACH ROW
+          EXECUTE PROCEDURE cdb_checkquota('0.001', '-1', 'public');
+      })
 
-        @user.tables.count.should eq 0
-        @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_false
+      @user.tables.count.should eq 0
+      @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_false
 
-        @ghost_tables_manager.link_ghost_tables_synchronously
+      @ghost_tables_manager.link_ghost_tables_synchronously
 
-        @user.tables.count.should eq 1
-        @user.tables.first.name.should == 'manolo_raster'
+      @user.tables.count.should eq 1
+      @user.tables.first.name.should == 'manolo_raster'
 
-        run_in_user_database(%{
-          DROP TABLE manolo_raster;
-        })
+      run_in_user_database(%{
+        DROP TABLE manolo_raster;
+      })
 
-        @user.tables.count.should eq 1
-        @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_false
-        @ghost_tables_manager.link_ghost_tables_synchronously
+      @user.tables.count.should eq 1
+      @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_false
+      @ghost_tables_manager.link_ghost_tables_synchronously
 
-        @user.tables.count.should eq 0
-        @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_true
-      end
+      @user.tables.count.should eq 0
+      @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_true
     end
 
     it 'should regenerate user tables with bad table_ids' do
